requirements.txt
django
djangorestframework
drf-yasg
psycopg2-binary
django-filter
djangorestframework-jwt
djangorestframework-simplejwt
django-cors-headers
httpx
django-json-widget
django-admin-rangefilter
ruff
pyotp
twilio
firebase-admin
paho-mqtt
